add options to flickity carousel. add block.json. create deprecated. separate save.js.
